Building fault detection and isolation related objects

FDIModel Falt detection and isolation synthesis model.

FDFilter Fault detection filter object.

FDFilterIF Fault detection filter internal form object.

fdimodset Setup of synthesis models for solving fault detection and isolation problems.

FaultDetectionTools.FDIModel — Type

FDIModel <: AbstractDescriptorStateSpace

Type for synthesis models employed to solve fault detection and isolation problems.

If sysf::FDIModel is the synthesis model object, the underlying descriptor system model can be obtained via sysf.sys and the indices of control, disturbance, fault, noise and auxiliary inputs can be accessed as the integer vectors contained in sysf.controls, sysf.disturbances, sysf.faults, sysf.noise and sysf.aux, respectively.

FaultDetectionTools.fdimodset — Function

fdimodset(sys; controls, c, disturbances, d, faults, f, fa, faults_sen, fs, noise, n, aux) -> sysf::FDIModel

Build for a given linear time-invariant system model sys = (A-λE,B,C,D), a synthesis model object sysf to be used in conjunction with the analysis and synthesis functions of FDI filters. The information on the partition of the input components in control, disturbance, fault, noise and auxiliary inputs can be specified using the following keyword arguments:

controls = inpu or c = inpu specifies the indices inpu of the control inputs (default: void)

disturbances = inpd or d = inpd specifies the indices inpd of the disturbance inputs (default: void)

faults = inpf or f = inpf specifies the indices inpf of the fault inputs (default: void)

fa = inpfa specifies the indices inpfa of control inputs subject to actuator fault (default: void)

faults_sen = inpfs or fs = inpfs specifies the indices inpfs of the system outputs subject to sensor fault inputs (default: void)

noise = inpn or noise = inpn specifies the indices inpn of the noise inputs (default: void)

aux = inpa specifies the indices inpa of the auxiliary inputs (default: void)

The indices of inputs or outputs can be specified as integer vectors, integer scalars or integer UnitRanges.

The resulting sysf contains the partitioned system sysf.sys = (A-λE,[Bu Bd Bf Bw Bv],C,[Du Dd Df Dw Dv]), where Bu, Bd, Bf, Bw and Bv are the input matrices from the control inputs u, disturbance inputs d, fault inputs f, noise inputs w and auxiliary inputs v, respectively, and Du, Dd, Df, Dw and Dv are the feedthrough matrices from those inputs. The indices of control, disturbance, fault, noise and auxiliary inputs are contained in the associated integer vectors sysf.controls, sysf.disturbances, sysf.faults, sysf.noise and sysf.aux, respectively.

Method: If G(λ) is the p x m transfer function matrix of sys, then the resulting system sysf has an equivalent input output form [Gu(λ) Gd(λ) Gf(λ) Gw(λ) Gv(λ)], where the following relations define the component matrices: Gu(λ) = G(λ)*Su, Gd(λ) = G(λ)*Sd, Gf(λ) = [G(λ)*Sf Ss], Gw(λ) = G(λ)*Sw, Gv(λ) = G(λ)*Sv, with the selection matrices Su, Sd, Sf, Sw and Sv formed from the columns of the m-th order identity matrix and Ss is formed from the columns of the p-th order identity matrix.

FaultDetectionTools.FDFilter — Type

FDFilter <: AbstractDescriptorStateSpace

Type for fault detection filters resulted as solutions of fault detection problems.

If filter::FDFilter is the fault detection filter object, the underlying descriptor system model can be obtained via filter.sys and the indices of output and control inputs can be accessed as the integer vectors contained in filter.outputs and filter.controls, respectively.

FaultDetectionTools.FDFilter — Method

FDFilter(sys, p, mu) -> Q::FDFilter

Build for a given linear time-invariant descriptor system model sys = (A-λE,B,C,D), a fault detection filter object Q, as determined with the synthesis functions of FDI filters. p and mu are the number of measured outputs and the number of control inputs, respectively. It is assumed that B = [By Bu Bv] and D = [Dy Du Dv] are partitioned matrices such that By and Dy have p columns, and Bu and Du have mu columns, where By and Bu are the input matrices from the measured outputs y and control inputs u, Dy and Du are the feedthrough matrices from the measured outputs y and control inputs u.

The resulting Q contains the partitioned system Q.sys = (A-λE,[By Bd],C,[Dy Du]) and the indices of inputs corresponding to the measured outputs and control inputs are contained in the associated integer vectors Q.outputs and Q.controls.

FaultDetectionTools.FDFilterIF — Type

FDFilterIF <: AbstractDescriptorStateSpace

Type for the internal form of fault detection filters resulted as solutions of fault detection problems.

If filter::FDFilterIF is the fault detection filter internal form object, the underlying descriptor system model can be obtained via filter.sys and the indices of control, disturbance, fault, noise and auxiliary inputs can be accessed as the integer vectors contained in filter.controls, filter.disturbances, filter.faults, filter.noise and filter.aux, respectively.

FaultDetectionTools.FDFilterIF — Method

FDFilterIF(sys, mu, md, mf, mw = 0, maux = 0; moff = 0 ) -> R::FDFilterIF

Build for a given linear time-invariant descriptor system model sys = (A-λE,B,C,D), a fault detection filter internal form object R, as determined with the synthesis functions of FDI filters. mu, md, mf, mw and maux are the dimensions of control, disturbance, fault, noise and auxiliary input vectors, respectively. It is assumed that B = [Boff Bu Bd Bf Bw Bv] and D = [Doff Du Dd Df Dw Dv] are partitioned matrices such that Boff and Doff have moff columns, Bu and Du have mu columns, Bd and Dd have md columns, Bf and Df have mf columns, Bw and Dw have mw columns, and Bv and Dv have maux columns.

The resulting R contains the partitioned system R.sys = (A-λE,[Bu Bd Bf Bw Bv],C,[Du Dd Df Dw Dv]) and the indices of inputs corresponding to the control, disturbance, fault, noise and auxiliary inputs are contained in the associated integer vectors R.controls, R.disturbances, R.faults, R.noise and R.aux.