Performance evaluation of FDI filters

fditspec Computation of the weak or strong structure matrix.

fdisspec Computation of the strong structure matrix.

fdscond Computation of the fault detection sensitivity condition.

FaultDetectionTools.fditspec — Function

S = fditspec(sysrf; FDfreq = missing, block = false, FDtol, FDStol, 
                    atol = 0, atol1 = atol, atol2 = atol, rtol, fast = true)

Compute the weak or strong binary structure matrix S of the transfer function matrix of a linear time-invariant system sysrf (typically representing the transfer channel from the fault inputs to residuals). sysrf is either a descriptor system representation sysrf = (Af-lambda*Ef,Bf,Cf,Df) with a q x mf transfer function matrix Rf(λ) or can be a fault detection filter internal form object sysrf::FDFilterIF, in which case only the fault inputs channel sysrf.sys[:,sysrf.faults] := (Af-lambda*Ef,Bf,Cf,Df) is selected.

If FDfreq = missing (default), then S contains the weak structure matrix of Rf(λ). For block = false, S is determined as a q x mf binary matrix (BitMatrix), whose (i,j)-th element is S[i,j] = 1, if the (i,j)-th element of Rf(λ) is nonzero, and otherwise, S[i,j] = 0. For block = true, S is determined as a 1 x mf binary matrix, whose (1,j)-th element is S[1,j] = 1, if the j-th column of Rf(λ) is nonzero, and otherwise, S[1,j] = 0.

If FDfreq = freq specifies a vector freq of nf real frequencies which characterize the classes of persistent fault signals, then S contains the strong structure matrix of Rf(λ) with respect to a set of nf complex frequencies Ω, defined as follows: if f is a real frequency in freq, then the corresponding complex frequency in Ω is λ := im*f, for a continuous-time system sysrf, or λ := exp(im*f*abs(Ts)), for a discrete-time system sysfr with sampling-time Ts. For block = false, S is determined as a q x mf x nf binary matrix, whose (i,j,k)-th element is S[i,j,k] = 1, if the (i,j)-th element of Rf(λ) is nonzero for the k-th frequency in freq, and otherwise, S[i,j,k] = 0. For block = true, S is determined as a 1 x mf x nf binary matrix, whose (1,j,k)-th element is S[1,j,k] = 1, if the j-th column of Rf(λ) is nonzero for the k-th frequency in FDfreq, and otherwise, S[1,j,k] = 0.

The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of matrices Af, Bf, Cf, Df, the absolute tolerance for the nonzero elements of Ef, and the relative tolerance for the nonzero elements of Af, Bf, Cf, Df and Ef. The keyword argument atol can be used to simultaneously set atol1 = atol and atol2 = atol.

FDtol = tol1 specifies an absolute threshold tol1 for the magnitudes of nonzero elements in the system matrices Bf and Df and is used to determine the weak structure matrix. Its default value is tol1 = 0.0001*max(1, norm(Bf,1), norm(Df,1)).

FDStol = tol2 specifies an absolute threshold tol2 for the magnitudes of nonzero elements in the system matrices Af, Ef, Bf, Cf and Df and is used to determine the strong structure matrix. Its default value is tol2 = epsm*max(1, norm(Ef,1), norm(Af,1), norm(Bf,1), norm(Cf,Inf), norm(Df,1))), where epsm is the working machine precision.

Method: For the definition of the structure matrix, see [1]. For the determination of the weak structure matrix, minimal realizations are determined for each column of Rf(λ) if block = true or for each element of Rf(λ) if block = false and the nonzero columns or elements in each column are identified (see Corollary 7.1 of [1]). For the determination of the strong structure matrix, minimal realizations are determined for each column of Rf(λ) if block = true or for each element of Rf(λ) if block = false and the full rank of the corresponding system matrix is checked for all frequencies in FDfreq (see Corollary 7.2 in [1]) (i.e., the lack of zeros in all frequencies).

References:

[1] Varga A. Solving Fault Diagnosis Problems - Linear Synthesis Techniques. Springer Verlag, 2017; sec.3.4.

FaultDetectionTools.fdisspec — Function

 fdisspec(sysrf, freq; block = false, FDGainTol = 0.01, 
                 atol = 0, atol1 = 0, atol2 = 0, rtol = 0, fast = true) -> (S, gains)

Compute for the linear time-invariant system sysrf and a given set of real frequencies freq, the strong binary structure matrix S of the transfer function matrix of sysrf and the corresponding frequency response gains gains. sysrf typically represents the transfer channel from the fault inputs to residuals and is either a descriptor system representation sysrf = (Af-lambda*Ef,Bf,Cf,Df) with a q x mf transfer function matrix Rf(λ) or is a fault detection filter internal form object sysrf::FDFilterIF, in which case only the fault inputs channel sysrf.sys[:,sysrf.faults] := (Af-lambda*Ef,Bf,Cf,Df) is selected.

freq must contain a real frequency value or a vector of nf real frequencies which characterize the classes of persistent fault signals (default: freq = 0, i.e., characterizing constant faults). S contains the strong structure matrix of Rf(λ) with respect to a set of nf complex frequencies Ω, defined as follows: if f is a real frequency in freq, then the corresponding complex frequency in Ω is λ := im*f, for a continuous-time system sysrf, or λ := exp(im*f*abs(Ts)), for a discrete-time system sysfr with sampling-time Ts. If any of the frequency values in freq is a pole of sysf, then sysf is replaced by the stable numerator of a left coprime factorization of sysf.

FDGainTol = tol specifies an absolute threshold tol for the nonzero magnitudes of the frequency response gains (default: tol = 0.01).

For block = false, gains is a p x mf x nf matrix, whose (i,j,k)-th element is the magnitude of the (i,j)-th element of Rf(λ) evaluated for the k-th frequency in freq. S is determined as a q x mf x nf binary matrix, whose (i,j,k)-th element is S[i,j,k] = 1, if the (i,j,k)-th element of gains is is larger then or equal to tol, and otherwise, S[i,j,k] = 0. For block = true, gains is an 1 x mf x nf matrix, whose (1,j,k)-th element is the norm of the j-th column of Rf(λ) evaluated for the k-th frequency in freq. S is determined as an 1 x mf x nf binary matrix, whose (1,j,k)-th element is S[1,j,k] = 1, if the (1,j,k)-th element of gains is larger then or equal to tol and otherwise, S[1,j,k] = 0.

The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of matrices Af, Bf, Cf, Df, the absolute tolerance for the nonzero elements of Ef, and the relative tolerance for the nonzero elements of Af, Bf, Cf, Df and Ef. The keyword argument atol can be used to simultaneously set atol1 = atol and atol2 = atol.

The computation of minimal realizations of individual input-output channels relies on pencil manipulation algorithms, which employ rank determinations based on either the use of rank revealing QR-decomposition with column pivoting, if fast = true, or the SVD-decomposition. The rank decision based on the SVD-decomposition is generally more reliable, but the involved computational effort is higher.

Method: S is evaluated using the definition of the strong structure matrix in [1].

References:

[1] Varga A. Solving Fault Diagnosis Problems - Linear Synthesis Techniques. Springer Verlag, 2017; sec. 3.4.

FaultDetectionTools.fdscond — Function

 fdscond(sysf,freq) -> (scond, β, γ)

Compute for a stable descriptor system sysf = (A-λE,B,C,D) with the transfer function matrix Rf(λ), β - the H∞- index of Rf(λ), γ - the maximum of the columns norms of Rf(λ) and the fault detection sensitivity condition scond evaluated as scond := β/γ. If freq is a vector of real frequency values, then β and γ are evaluated over the frequencies contained in freq. If sysf is a fault detection filter internal form object sysf::FDFilterIF then scond, β and γ are evaluated only for the fault inputs channel sysf.sys[:,sysf.faults].