FDI related utilities

fdhinfminus Evaluation of the H∞- index of the transfer function matrix of a descriptor system model.

fdhinfmax Evaluation of the maximum of column norm of the transfer function matrix of a descriptor system model.

fditspec_ Computation of the weak or strong structure matrix of a descriptor system model.

fdisspec_ Computation of the strong structure matrix of a descriptor system model.

fdscond_ Computation of the column-gains sensitivity condition of the transfer function matrix of a descriptor system model.

FaultDetectionTools.fdhinfminus — Function

 fdhinfminus(sys,freq) -> (β, ind, fr)

Compute for a stable descriptor system sys = (A-λE,B,C,D) the H∞- index β of its transfer function matrix G(λ). If freq = missing (default), then β is the minimum H∞-norm of the columns of G, ind is the index of the minimum-norm column and fr is the frequency where the minimum H∞-norm of the columns is achieved. If freq is a real value or a real vector of frequency values, then β is the minimum of the 2-norms of the columns of the frequency responses of G evaluated for all values contained in freq, ind is the index of column for which the minimum is achieved and fr is the corresponding frequency.

FaultDetectionTools.fdhinfmax — Function

 fdhinfmax(sys,freq) -> (γ, ind, fr)

Compute for a descriptor system sys = (A-λE,B,C,D), γ - the maximum norm of the columns of its transfer function matrix G(λ). If freq = missing (default), then γ is the maximum H∞-norm of the columns of G, ind is the index of the maximum-norm column and fr is the frequency where the maximum H∞-norm of the columns is achieved. If freq is a real value or a real vector of frequency values, then γ is the maximum of the 2-norms of the columns of the frequency responses of G evaluated for all values contained in freq, ind is the index of column for which the maximum is achieved and fr is the corresponding frequency.

FaultDetectionTools.fditspec_ — Function

S = fditspec_(sysrf::DescriptorStateSpace; FDfreq = missing, block = false, poleshift = false, 
             FDtol, FDStol, atol = 0, atol1 = atol, atol2 = atol, rtol, fast = true)

Compute the weak or strong binary structure matrix S of the transfer function matrix of a linear time-invariant system sysrf (typically representing the transfer channel from the fault inputs to residuals). sysrf has a descriptor system realization of the form sysrf = (Af-lambda*Ef,Bf,Cf,Df) with a q x mf transfer function matrix Rf(λ). For the description of keyword parameters see the documentation of fditspec.

FaultDetectionTools.fdisspec_ — Function

 fdisspec_(sysrf::DescriptorStateSpace, freq; block = false, FDGainTol = 0.01, 
                 atol, atol1, atol2, atol3, rtol, fast = true) -> (S, gains)

Compute the strong binary structure matrix S of the transfer function matrix of a linear time-invariant system sysrf (typically representing the transfer channel from the fault inputs to residuals). sysrf has a descriptor system realization of the form sysrf = (Af-lambda*Ef,Bf,Cf,Df) with a q x mf transfer function matrix Rf(λ). For the description of keyword parameters see the documentation of fdisspec.

FaultDetectionTools.fdscond_ — Function

 fdscond_(sysrf::DescriptorStateSpace, freq) -> (scond, β, γ)

Compute for a stable descriptor system sysrf = (A-λE,B,C,D) with the transfer function matrix Rf(λ), β - the H∞- index of Rf(λ), γ - the maximum of the columns norms of Rf(λ) and scond - the column-gains sensitivity condition evaluated as scond := β/γ. If freq is a vector of real frequency values, then β and γ are evaluated over the frequencies contained in freq.